Require `View: ViewMarker` to avoid the leaky abstraction `fn view() -> impl View + ViewMarker`
Motivation: to avoid requiring `fn view() -> impl View + ViewMarker` and instead just using `fn view() -> impl View`

I don't see a reason for types that implement `View`, but not `ViewMarker`. This requirement has the neat side-effect of avoiding weird error-messages otherwise, in case `impl ViewMarker for T` has been forgotten for `impl View for T`.

The whole purpose AFAIK of `ViewMarker` is a workaround of the orphan rule, that something like this:
```rust
impl<T, A, V: View<T, A> + ViewMarker> ViewSequence<T, A> for V
```
is possible.

So I tried adding this as super trait requirement and it seems to be possible.
I think this makes the leaky abstraction/workaround much more feasible as it's only required now for actual `View` implementations.