xilem_web: Rewrite modifiers (`Attributes`, `Classes` and `Styles`)
Previously the modifier systems had design issues i.e. bugs (non-deleted styles/classes/attributes), and were unnecessary complex.
This aims to solve this by not using separate traits, but concrete types and a different mechanism that is closer to how `ElementSplice` works.

There's a few fundamental properties that composable type-based modifiers need to support to avoid surprising/buggy behavior:

* Minimize actual changes to the underlying element, as DOM traffic is expensive.
* Be compatible to memoization: e.g. a `Rotate` view should still be applicable to possibly memoized transform values of the underlying element.
* Recreation when the underlying element has changed (e.g. with a change of variants of a `OneOf`).

To support all this, the modifier system needs to retain a modifier for each modifier-view, and track its changes independently of the view.
Previously all elements were directly written and separated with markers, to limit the boundaries of such views, but this had issues, when e.g. all modifiers were deleted (e.g. clearing a `Vec` of classes), by not deleting these.

With this PR, the count of modifiers of a modifier-view is directly stored either in the view or its view state, which cleans up the concrete modifier elements, and makes it less prone for errors.

The API to use these modifiers in modifier-views was also redesigned to hopefully be more straight-forward/idiomatic.

One interesting aspect compared to the previous system is the use of a new trait `With` for modifiers.
Instead of (roughly) `Element: WithStyle`, it works with `Element: With<Styles>`.
This prevents all kinds of reimplementations of something like `WithStyle` for elements.
This gets especially visible in the `one_of` module, which can be covered by one blanket implementation.